company_brochure_ui/requirements.txt
python-dotenv
requests
openai
gradio